Add donate button. Fix 403 error
Sometimes the app would display a 403 page when the user resumed. This
is caused by the app being PLM'd when on the page displayed after
sending a text. If we're on this page, don't store the URL in PageState.

Also adds a button to donate.

Less of a crashing chance on the log-out piece.